merge branch develop into main

+++ Mecanillama.API/Appointments/Domain/Models/Appointment.cs
@@ -6,12 +6,12 @@ namespace Mecanillama.API.Appointments.Domain.Models;
 public class Appointment
 {
     //Properties
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Date { get; set; }
     public string Time { get; set; }
     //Relationships - many to many
-    public long CustomerId { get; set; }
-    public long MechanicId { get; set; }
+    public int CustomerId { get; set; }
+    public int MechanicId { get; set; }
     public Customer Customer { get; set; }
     public Mechanic Mechanic { get; set; }
